refactor(frontend): extract useResizablePanes composable (P2-2 F5)
Move the three-pane layout cluster (sidebar/right width ratios + clamps +
computed styles, left-pane heights/refs, the three mouse drag handlers, and
the window-resize listener lifecycle) out of DataStudioNew.vue into
composables/useResizablePanes.js. The composable owns its mousemove/mouseup
handler vars and registers its own onMounted/onBeforeUnmount for the resize
listener + drag cleanup; activeTab is injected and syncResultPaneLayout is
passed as a lazy wrapper (forward ref). Returned refs (leftPaneHeights/
leftPaneRefs) stay shared with disposeTabResources.

Verified: DataStudioNew mount smoke green, npm run build (refs resolve),
npm run lint (0 errors); orphan check shows all moved symbols gone from the
component. Drag/resize logic is verbatim-moved (behavior preserved by
construction); interactive drag itself is not exercised by the smoke.
Refs docs/plans/2026-06-18-frontend-megacomponent-split-plan.md (F5).
